third_party/WebKit/LayoutTests/shadow-dom/v0/deep-in-selectors.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id=a><div id=b></div></div>
<script>
test(() => {
  const sr = a.createShadowRoot();
  const p = document.createElement('p');
  sr.appendChild(p);

  assert_array_equals(document.querySelectorAll('#a *, #a /deep/ *'), [p, b]);
}, '/deep/ should be valid in the second selector of selectors');
</script>
